server/requirements.txt
flask==3.0.0
torch==2.6.0
transformers==4.35.2
python-dotenv==1.0.0
